Classifier/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-mock>=3.11.0
# Raspberry Pi only (install separately on Pi):
# picamera2>=0.3.12
# RPi.GPIO>=0.7.1
